pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Guilherme Chalita <guilhermechalita@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.1"
factory-boy = "^3.3.0"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pytest-django = "^4.5.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
